error checking

+++ setup.sh
@@ -241,9 +241,9 @@ EOF
     elif [ VARVAL_CHAIN_NAME == "mainnet" ]; then
         CHAIN_DIR="xdcchain"
     fi
-    
+
     echo
-    echo -e "${YELLOW}Chain Directory is:  $CHAIN_DIR.${NC}"
+    echo -e "${YELLOW}Chain Directory is:  $CHAIN_DIR${NC}"
     echo -e "${YELLOW}Correcting chain directory permissions $CHAIN_DIR.${NC}"
 
     sudo chown $USER_ID:$USER_ID -R $CHAIN_DIR
@@ -641,8 +641,17 @@ server {
 }
 EOF
     sudo chmod 644 $NGX_CONF_NEW
-    sudo ln -s $NGX_CONF_NEW /etc/nginx/sites-enabled/
-    sudo rm -f $NGX_CONF_OLD
+
+    #check if symbolic link file exists in sites-enabled
+    if [ ! -f /etc/nginx/sites-enabled/$NGX_CONF_NEW ]; then
+        sudo ln -s $NGX_CONF_NEW /etc/nginx/sites-enabled/
+    fi
+    
+    #delete default symbolic link file if it exists in sites-enabled
+    if [  -f /etc/nginx/sites-enabled/default ]; then
+        sudo rm -f /etc/nginx/sites-enabled/default
+    fi   
+    
     # Reload Nginx to apply the new configuration
     sudo systemctl reload nginx
